requirements.txt
django
pillow
pixelmatch
html2image